backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1
PyJWT==2.8.0

# File Upload & Storage
Pillow==10.2.0
django-storages==1.14.2
boto3==1.34.34

# Real-time Features
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0

# API Documentation
drf-spectacular==0.27.1

# Environment Variables
python-decouple==3.8
python-dotenv==1.0.1

# Geolocation & Maps
geopy==2.4.1
googlemaps==4.10.0

# Payment Integration
stripe==8.2.0

# WhatsApp & Telegram Bot
python-telegram-bot==20.7
twilio==8.11.1

# Email
django-ses==3.5.2

# Utilities
requests==2.31.0
celery==5.3.6
redis==5.0.1

# Testing
pytest==8.0.0
pytest-django==4.7.0
factory-boy==3.3.0

# Development
django-extensions==3.2.3
ipython==8.20.0
